Update submodule and commit after gem publish
This lets us set the jenkins job for this repo as a post commit hook to
the govuk_frontend_toolkit repo. It will ensure the gem is always up to
date with that repository.

Bumping the gem will then be a fully automatic process controlled by the
govuk_frontend_toolkit repo.