Remove gem dependency on sass
Ruby sass is now end-of-life and should not be used. It is somewhat
peculiar that this gem sets this as a dependency as this gem alone isn't
enough to integrate with Rails (that would need sass-rails or
sassc-rails) and this code itself does not depend on it.

I've chosen to remove this rather than switch this to be sass-rails or
sassc-rails as I expect all apps that use this legacy tool already have
their own sass compiler and adding one could make for another problematic
dependency. Since the installation guidance [1] is to add the SASS as
@import calls it seems implied that it should only function in
applications that already have SASS installed.

[1]: https://github.com/alphagov/govuk_frontend_toolkit_gem/tree/9bae66b74adc88af12c89b513e52b0c5b6f675de#usage